BUENAS Y MALAS NOTICIAS ¡La funcion crearArbolDecision crea un arbol! Pero solo si el numero de datos es pequeño... Por ahora genera un arbol con 50 elementos pero con 100 ya no... Hay que ver que coño le pasa... Yo ya no puedo mas por hoy, que ya son las 02:30 (-_-) Buenas noches

+++ lib/pilaDeadAlive.h
@@ -14,10 +14,10 @@
 
 typedef struct tipoElementoPila_t{
 	int pos;
-	int vivos_Izq;	
-	int vivos_Dch;	
-	int muertos_Izq;	
-	int muertos_Dch;	
+	double vivos_Izq;	
+	double vivos_Dch;	
+	double muertos_Izq;	
+	double muertos_Dch;	
 }tipoElementoPila;
 typedef struct celdaP{
 	 tipoElementoPila elem;
